When a user create a new account, he'll be automatically logged in the system

+++ server/src/main/java/ch/waterbead/services/ReservationService.java
@@ -1,10 +1,16 @@
 package ch.waterbead.services;
 
+import java.time.LocalDate;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import ch.waterbead.models.Reservation;
+import ch.waterbead.models.User;
 import ch.waterbead.repositories.ReservationRepository;
+import ch.waterbead.util.Response;
 
 @Service
 public class ReservationService {
@@ -12,14 +18,52 @@ public class ReservationService {
 	ReservationRepository reservationRepository;
 	
 	public void add(Reservation reservation) {
-		reservationRepository.saveAndFlush(reservation);
+		User user = getCurrentUser();
+		reservation.setUser(user);
+		if(isReservationsNotExist(reservation)) {
+			reservationRepository.saveAndFlush(reservation);
+		}
 	}
 	
 	public void update(Reservation reservation) {
-		reservationRepository.saveAndFlush(reservation);
+		User user = getCurrentUser();
+		reservation.setUser(user);
+		if(hasRightToModify(reservation) && !isReservationsNotExistForUpdate(reservation)) {
+			reservationRepository.saveAndFlush(reservation);
+		}
 	}
 	
-	public void delete(Reservation reservation) {
-		reservationRepository.delete(reservation);
+	public void deleteById(Long id) {
+		Reservation reservation = reservationRepository.findOne(id);
+		User user = getCurrentUser();
+		reservation.setUser(user);
+		if(hasRightToModify(reservation)) {
+			reservationRepository.delete(reservation);
+		}
+	}
+	
+	private boolean isReservationsNotExist(Reservation reservation) {
+		return reservationRepository.findReservationsChevauchantes(reservation).size() == 0;
+	}
+	
+	private boolean isReservationsNotExistForUpdate(Reservation currentReservation) {
+		List<Reservation> reservations = reservationRepository.findReservationsChevauchantes(currentReservation);
+		if(reservations.size() == 0) return false;
+		if(reservations.size() > 1) return true;
+		if(reservations.get(0).equals(currentReservation)) return false;
+		return true;
+	}
+	
+	private boolean hasRightToModify(Reservation reservation) {
+		User user = getCurrentUser();
+		Reservation reservationExistante = reservationRepository.findOne(reservation.getId());
+		if(reservationExistante.getUser().equals(user)) {
+			return true;
+		}
+		return false;
+	}
+	
+	private User getCurrentUser() {
+		return (User)SecurityContextHolder.getContext().getAuthentication().getPrincipal();
 	}
 }
